comment voting extracted to separate module
