+++ server/src/modules/comment-voting/Resolvers.ts
@@ -1,0 +1,27 @@
+import { container } from "@/inversify.config";
+import { CommentVoteService } from "./index";
+import { Unauthorized } from "@/exceptions/exceptions";
+import { AuthProvider, IdToken } from "@/modules/auth";
+
+const authProvider = container.get(AuthProvider);
+const service = container.get(CommentVoteService);
+
+export const commentVoteResolvers = {
+  Mutation: {
+    removeCommentVote: async (_, args, { idToken }: { idToken: IdToken }) => {
+      if (!idToken) throw Unauthorized("You have to be authorized to vote.");
+      const decodedToken = await authProvider.verifyIdToken(idToken, true);
+      return await service.removeVote(args.commentId, decodedToken.uid);
+    },
+    addCommentVoteUp: async (_, args, { idToken }: { idToken: IdToken }) => {
+      if (!idToken) throw Unauthorized("You have to be authorized to vote.");
+      const decodedToken = await authProvider.verifyIdToken(idToken, true);
+      return await service.setVoteUp(args.commentId, decodedToken.uid);
+    },
+    addCommentVoteDown: async (_, args, { idToken }: { idToken: IdToken }) => {
+      if (!idToken) throw Unauthorized("You have to be authorized to vote.");
+      const decodedToken = await authProvider.verifyIdToken(idToken, true);
+      return await service.setVoteDown(args.commentId, decodedToken.uid);
+    },
+  },
+};
