+++ server/src/modules/comment-voting/Schemas.ts
@@ -1,0 +1,9 @@
+import { gql } from "apollo-server-micro";
+
+export const commentVoteTypeDefs = gql`
+  extend type Mutation {
+    removeCommentVote(commentId: ID!): Boolean!
+    addCommentVoteUp(commentId: ID!): Boolean!
+    addCommentVoteDown(commentId: ID!): Boolean!
+  }
+`;
