+++ server/src/modules/comments/Schemas.ts
@@ -41,8 +41,5 @@ export const commentsTypeDefs = gql`
 
   extend type Mutation {
     createComment(comment: CommentInput!): Comment!
-    removeCommentVote(commentId: ID!): Boolean!
-    addCommentVoteUp(commentId: ID!): Boolean!
-    addCommentVoteDown(commentId: ID!): Boolean!
   }
 `;
